Separate top level test groups into files
